libTUStereo++ added. testv4l2cmaera updated. iterator.h updated.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@1611 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUV4L2++/testv4l2camera/MyV4L2Camera.h
@@ -33,7 +33,7 @@ class MyV4L2Camera : public V4L2Camera
   //! 画像の表示領域となるキャンバスを返す．
     GtkWidget*		canvas()				const	;
     V4L2Camera&		setFormat(PixelFormat pixelFormat,
-				  u_int width, u_int height,
+				  size_t width, size_t height,
 				  u_int fps_n, u_int fps_d)		;
     void		idle()						;
     void		draw()						;
